+ refactor login logic + added Photo.java
