+++ src/main/java/com/example/sp/Photo/Photo.java
@@ -1,0 +1,4 @@
+package com.example.sp.Photo;
+
+public class Photo {
+}
