+++ src/main/java/com/example/sp/SpProject2021Application.java
@@ -1,14 +1,10 @@
 package com.example.sp;
 
-import com.example.sp.model.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import com.example.sp.repository.UserRepository;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import com.example.sp.User.repository.UserRepository;
 
 @SpringBootApplication
 public class SpProject2021Application implements CommandLineRunner {
